requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain_community
langchain_openai
langchain_pinecone
langchain_experimental
-e .